# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Badminton Project API",
        "description": "Badminton Project API Documentation",
        "contact": {
            "email": "admin@badminton.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Local Server"
        },
        {
            "url": "https://api.badminton.example.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/bank-account/{userId}": {
            "get": {
                "tags": [
                    "BankAccount"
                ],
                "summary": "Lấy thông tin ngân hàng của người dùng",
                "description": "Lấy thông tin ngân hàng dựa vào ID người dùng",
                "operationId": "getUserBankAccount",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID của người dùng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thông tin thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy thông tin ngân hàng thành công"
                                        },
                                        "bankAccount": {
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy người dùng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Người dùng không tồn tại"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}